+++ frontend/src/pages/HomePage.jsx
@@ -3,7 +3,7 @@ import dabbingBeaverImg from "../assets/dabbingBeaver.svg";
 import abigailImg from "../assets/abigail.svg";
 import ramonaImg from "../assets/ramona.svg";
 
-function HomePage() {
+export default function HomePage() {
   return (
     <>
       <main className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
@@ -188,5 +188,3 @@ function HomePage() {
     </>
   );
 }
-
-export default HomePage;